wip: Moved lang selector to layout with localstorage caching
